requirements.txt
apache-airflow==2.7.2
transformers==4.34.0
pandas==1.5.3
beautifulsoup4==4.12.2
requests==2.31.0

